MazF from Zhang et al. 2019

In the following, the use of of the rt-arrest function will be applied to a MazF digestion assay from Zhang et al. 2019. Herein, 3 replicates of HEK293 mRNA were treated with FTO or mock treated and then subjected to a MazF digestion assay.

data(MazF)

Format

a tibble with 22 elements:

id: Character string representing a unique identifier - created from contig, start, [end], and strand.

contig: Character string representing the contig of the variant

start: Numeric position of variant (>=0)

end: Numeric corresponds to "start + 1"

name: Character string. Name of used method call-2

pvalue: Numeric value representing the pvalue of the test.

strand: Character representing strand information; "+", "-", or "."(no strand information available)

arrest: Numeric tibble with representing counts for A, C, G, and T base calls from arrest reads.

through: Numeric tibble with representing counts for A, C, G, and T base calls from through reads.

bases: Numeric tibble with representing counts for A, C, G, and T base calls.

cov: Numeric value indicating the read coverage for this site

arrest_rate: Numeric tibble representing the arrest rate for each sample.

arrest_score: Numeric - test-statistic score.

backtrack1: Character - indicator if backtracking was used for condition 1.

backtrack2: Character - indicator if backtracking was used for condition 2.

backtrackP: Character - indicator if backtracking was used for condition pooled condition.

reset1: Character - indicator if default estimation was unstable with for condition 1.

reset2: Character - indicator if default estimation was unstable with for condition 2.

resetP: Character - indicator if default estimation was unstable with for pooled condition.

info: Character string separated with ";" provding additional data for this specific site. Empty field is equal to "*"

filter: ";"-separated character string showing feature filter information. Empty field is equal to "*"

ref: Character "A", "C", "G", "T", or "N" representing the reference base for this site - inverted when strand is "-".

References

Zhang, Zhang and Chen, Li-Qian and Zhao, Yu-Li and Yang, Cai-Guang and Roundtree, Ian A. and Zhang, Zijie and Ren, Jian and Xie, Wei and He, Chuan and Luo, Guan-Zheng Science advances, July 5, 2019