Filters sites by artefacts.

Removes sites that sites that have been marked by feature/artefact filter.

filter_artefact(filter, artefacts = NULL)

Arguments

filter
vector of strings that contains artefact filter information.

artefacts
vector of characters that correspond to feature/artefact filters to be filtered, default: NULL - Filter all.

Value

vector of logical.

Examples

data(rdd)
# remove sites that are marked by artefact filter "D"
dim(rdd)

#> [1] 94 15

dim(dplyr::filter(rdd, filter_artefact(filter, c("D"))))

#> [1] 27 15

# remove all sites that are marked by some artefact filter
dim(dplyr::filter(rdd, filter_artefact(filter)))

#> [1] 62 15